Fix #195. AbstractButton no longer implements HasHref and HasTargetHistoryToken. AnchorButton now implements HasHref and HasTargetHistoryToken. AnchorButton's target history token handling is improved based on Hyperlink's implementation.
