+++ gwtbootstrap3/src/main/java/org/gwtbootstrap3/client/ui/AnchorButton.java
@@ -20,10 +20,16 @@ package org.gwtbootstrap3.client.ui;
  * #L%
  */
 
+import com.google.gwt.core.client.GWT;
 import com.google.gwt.dom.client.AnchorElement;
 import com.google.gwt.dom.client.Document;
 import com.google.gwt.dom.client.Element;
+import com.google.gwt.user.client.DOM;
+import com.google.gwt.user.client.Event;
+import com.google.gwt.user.client.History;
+import com.google.gwt.user.client.ui.impl.HyperlinkImpl;
 import org.gwtbootstrap3.client.ui.base.HasHref;
+import org.gwtbootstrap3.client.ui.base.HasTargetHistoryToken;
 import org.gwtbootstrap3.client.ui.base.button.AbstractToggleButton;
 import org.gwtbootstrap3.client.ui.constants.ButtonType;
 
@@ -34,19 +40,66 @@ import org.gwtbootstrap3.client.ui.constants.ButtonType;
  * @see Button
  * @see org.gwtbootstrap3.client.ui.base.button.AbstractToggleButton
  */
-public class AnchorButton extends AbstractToggleButton implements HasHref {
+public class AnchorButton extends AbstractToggleButton implements HasHref, HasTargetHistoryToken {
+
+    private String targetHistoryToken;
+
+    private static HyperlinkImpl impl = GWT.create(HyperlinkImpl.class);
 
     public AnchorButton(final ButtonType type) {
         super(type);
         setHref(EMPTY_HREF);
+        sinkEvents(Event.ONCLICK);
+    }
+
+    @Override
+    public void onBrowserEvent(Event event) {
+        super.onBrowserEvent(event);
+        if (getTargetHistoryToken() != null) {
+            // implementation is based on Hyperlink#onBrowserEvent
+            if (DOM.eventGetType(event) == Event.ONCLICK && impl.handleAsClick(event)) {
+                History.newItem(getTargetHistoryToken());
+                event.preventDefault();
+            }
+        }
+    }
+
+    /**
+     * Set the target history token for the widget. Note, that you should use either {@link #setTargetHistoryToken(String)}
+     * or {@link #setHref(String)}, but not both as {@link #setHref(String)} resets the target history token.
+     * @param targetHistoryToken String target history token of the widget
+     */
+    @Override
+    public void setTargetHistoryToken(final String targetHistoryToken) {
+        this.targetHistoryToken = targetHistoryToken;
+        if (targetHistoryToken != null) {
+            final String hash = History.encodeHistoryToken(targetHistoryToken);
+            getAnchorElement().setHref("#" + hash);
+        }
+    }
+
+    /**
+     * Get the target history token for the widget. May return {@code null} if no history token has been set or if
+     * it has been reset by {@link #setHref(String)}
+     * @return String the widget's target history token
+     */
+    @Override
+    public String getTargetHistoryToken() {
+        return targetHistoryToken;
     }
 
     public AnchorButton() {
         this(ButtonType.DEFAULT);
     }
 
+    /**
+     * Set's the HREF of the widget. Note, that you should use either {@link #setTargetHistoryToken(String)}
+     * or {@link #setHref(String)}, but not both as {@link #setHref(String)} resets the target history token.
+     * @param href String href
+     */
     @Override
     public void setHref(final String href) {
+        this.targetHistoryToken = null;
         getAnchorElement().setHref(href);
     }
